requirements.txt
aiogram==3.1.0
psycopg2-binary
python-dotenv
